Move query handling functions to separate file for better separation of concerns
